pyproject.toml
[tool.poetry]
name = "stagereminder"
version = "0.1.0"
description = ""
authors = ["GStarP"]
packages = [
    { include = "stagereminder", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.27.2"
qq-botpy = "^1.2.1"
python-dotenv = "^1.0.1"
openai = "^1.55.0"
sqlalchemy = "^2.0.36"
apscheduler = "^3.11.0"
pytz = "^2024.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"

[tool.pytest.ini_options]
addopts = "-v -s"
pythonpath = ["src"]
testpaths = ["src/stagereminder"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
